core: Do not store opdata array in flash
ESP8266 and ESP32 do not have internal flash, but rely on an external
flash chip. This makes reading and writing to from flash extremely slow
and is advised not to do this [0]. For ESP32 this is actually even a
dummy operation.

For ESP it's a backwards compatibility thing, as it's more of an AVR
thing.

[0]: https://arduino.stackexchange.com/questions/71250/use-of-progmem-in-esp8266-vs-avr-and-also-how-to-handle-large-dynamic-strings

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>