core: Increase readability of output check
When comparing the byte to the bitmask we know it won't be zero, but
checking whether it is 'greater then' zero is cognitively harder.

Why larger then zero? we're just checking if a bit matches, why not
compare against '1', oh the bitmask shifts of course, why are we not
comparing against bitmask (also an option).

The other way around is easier to read and just as logical. If the bit
is zero, we write 0, otherwise 1. Technically identical, much easier to
read however.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>