core: Always generate extended miso data
While we don't have to generate the extended frame data for standard
transactions, it won't hurt to generate them anyway. The code overhead
is tiny, and for platforms that do support extended frames, we gotta be
able to support the CPU load anyway, so lets always prepare the frame.

When sending, we do take frame size into account, so the extra bits
aren't sent anyway.

This also opens the path for frame size auto detection, either based on
signature (unconfirmed) or by just parsing the data (if we parse an
extended frame, and we get checksum errors on the extended frame, but
not on the standard frame, we know what we are. Though that is future
thoughts.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>